Check ibus emoji picker keybinding before overwriting
The current implementation will overwrite the existing ibus emoji picker keybinding regardless whether or not the existing keybinding is the same as the desired keybinding.

This commit adds a check before overwriting the Ibus emoji picker keybinding.

It might be wiser to write my own module for handling gsettings commands. This will clean up the output a bit and make it much cleaner to handle any future use of gsettings commands.